requirements.txt
flask
gunicorn
numpy
pillow
opencv-python
torch
torchvision
tensorflow
scikit-image
git-lfs
